btlFormation: 3 over-window repairs (002BA3A0 1396->1376 nd543->444, 002BCE10 604->592 nd368->332, 002BB3A0 484->480 nd324->224)

+++ src/Event/Fcl/fclCombineList.c
@@ -2406,7 +2406,7 @@ u64 FUN_003df4a0(u64 param_1,int param_2)
 
   int iVar1;
 
-  short sVar2;
+  s32 sVar2;
 
   float fVar3;
